Fixed inaccurate double->float conversion (volume)

+++ CSC8503/CSC8503Common/AudioSystem.h
@@ -76,7 +76,7 @@ namespace NCL {
 
 			//global settings
 			static void SetGlobalVolume(float vol) { vol = min(vol, 1.0f); vol = max(vol, 0.0f); engine->setSoundVolume(vol); if (s) s->SetVolume(vol); }
-			static void SetGlobalVolume(double vol) { vol = min(vol, 1.0); vol = max(vol, 0.0); engine->setSoundVolume(vol); if (s) s->SetVolume(vol); }
+			static void SetGlobalVolume(double vol) { vol = min(vol, 1.0); vol = max(vol, 0.0); engine->setSoundVolume((float)vol); if (s) s->SetVolume(vol); }
 			static float GetGlobalVolume() { return (float)engine->getSoundVolume(); }
 
 			static void SetDefaultMinDistance(float d) { engine->setDefault3DSoundMinDistance(d); }
